requirements.txt
tensorflow-gpu>=2.0.3
tqdm
pandas
glob3
matplotlib
